make an interaction with ui

+++ entity-relation/src/model/Entity.java
@@ -2,15 +2,15 @@ package model;
 
 public class Entity {
 
-    private Long id_E;
+    private Integer id_E;
     private String name_E;
 
-    public Entity(Long id_e, String name_e) {
+    public Entity(Integer id_e, String name_e) {
         id_E = id_e;
         name_E = name_e;
     }
 
-    public void setId_E(Long id_E) {
+    public void setId_E(Integer id_E) {
         this.id_E = id_E;
     }
 
@@ -18,7 +18,7 @@ public class Entity {
         this.name_E = name_E;
     }
 
-    public Long getId_E() {
+    public Integer getId_E() {
         return id_E;
     }
