+++ entity-relation/src/model/Entity_Attribute.java
@@ -1,22 +1,23 @@
 package model;
 
 public class Entity_Attribute {
-    private Long id_EA;
+
+    private Integer id_EA;
     private String name_EA;
 
-    private Long id_E;
+    private Integer id_E;
 
-    public Entity_Attribute(Long id_EA, String name_ea, Long id_E) {
+    public Entity_Attribute(Integer id_EA, String name_ea, Integer id_E) {
         this.id_E = id_E;
         this.id_EA = id_EA;
         this.name_EA = name_ea;
     }
 
-    public Long getId_E() {
+    public Integer getId_E() {
         return id_E;
     }
 
-    public Long getId_EA() {
+    public Integer getId_EA() {
         return id_EA;
     }
 
@@ -24,11 +25,11 @@ public class Entity_Attribute {
         return name_EA;
     }
 
-    public void setId_E(Long id_E) {
+    public void setId_E(Integer id_E) {
         this.id_E = id_E;
     }
 
-    public void setId_EA(Long id_EA) {
+    public void setId_EA(Integer id_EA) {
         this.id_EA = id_EA;
     }
